refactor & bugfix

+++ rtmp-sharp-server-core/Controller/LivingController.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Security.Authentication;
+using System.Threading;
 using System.Threading.Tasks;
 using Complete;
 using RtmpSharp.IO;
@@ -47,6 +48,12 @@ namespace RtmpSharp.Controller
 
         private SessionStorage _sessionStorage = null;
         private PublisherSessionService _publisherSessionService = null;
+        private int _streamId = 0;
+        private Random _random = new Random();
+
+        private const int COMMAND_CHANNEL = 4;
+        private const int VIDEO_CHANNEL = 5;
+        private const int AUDIO_CHANNEL = 6;
 
         public LivingController(PublisherSessionService publisherSessionService)
         {
@@ -70,7 +77,7 @@ namespace RtmpSharp.Controller
         }
 
         [RpcMethod(Name = "publish")]
-        public object Publish(string publishingName, string publishingType)
+        public void Publish(string publishingName, string publishingType)
         {
             var publishingTypeMap = new Dictionary<string, PublishingType>()
             {
@@ -89,15 +96,13 @@ namespace RtmpSharp.Controller
 
             _sessionStorage.PublishingType = publishingTypeMap[publishingType];
 
-            _sessionStorage.AudioBuffer = new Queue<AudioData>();
-            _sessionStorage.VideoBuffer = new Queue<VideoData>();
             _publisherSessionService.RegisterPublisher(publishingName, Session);
             Session.Disconnected += (s, e) =>
             {
                 _publisherSessionService.RemovePublisher(Session);
             };
 
-            Session.WriteProtocolControlMessage(new UserControlMessage(UserControlMessageType.StreamBegin, new int[] { Session.StreamId }));
+            Session.WriteProtocolControlMessage(new UserControlMessage(UserControlMessageType.StreamBegin, new int[] { _streamId }));
 
             Session.NotifyStatus(new AsObject
             {
@@ -105,23 +110,22 @@ namespace RtmpSharp.Controller
                 {"code", "NetStream.Publish.Start" },
                 {"description", "Stream is now published." },
                 {"details", publishingName }
-            });
+            }, _streamId, COMMAND_CHANNEL);
 
             _sessionStorage.IsPublishing = true;
-            return new object();
         }
 
         public void Dispose()
         {
-            if (Session != null)
+            if (Session != null && Session.SessionStorage != null)
             {
-                if (_sessionStorage.ConnectedSession != null)
+                if (Session.SessionStorage.ConnectedSession != null)
                 {
-                    var publisherSessionStorage = _sessionStorage.ConnectedSession.SessionStorage as SessionStorage;
+                    var publisherSessionStorage = Session.SessionStorage.ConnectedSession.SessionStorage as SessionStorage;
                     if (publisherSessionStorage != null)
                     {
-                        publisherSessionStorage.VideoReceived -= SendVideoData;
-                        publisherSessionStorage.AudioReceived -= SendAudioData;
+                        publisherSessionStorage.VideoReceived -= BufferVideoData;
+                        publisherSessionStorage.AudioReceived -= BufferAudioData;
                     }
                     _publisherSessionService.RemovePublisher(Session);
                 }
@@ -136,7 +140,8 @@ namespace RtmpSharp.Controller
                 throw new UnauthorizedAccessException();
             }
 
-            Session.WriteProtocolControlMessage(new UserControlMessage(UserControlMessageType.StreamBegin, new int[] { Session.StreamId }));
+            Session.WriteProtocolControlMessage(new UserControlMessage(UserControlMessageType.StreamIsRecorded, new int[] { _streamId }));
+
 
             Session.NotifyStatus(new AsObject
             {
@@ -144,14 +149,14 @@ namespace RtmpSharp.Controller
                 {"code", "NetStream.Play.Reset" },
                 {"description", "Resetting and playing stream." },
                 {"details", streamName }
-            });
+            }, _streamId, COMMAND_CHANNEL);
             Session.NotifyStatus(new AsObject
             {
                 {"level", "status" },
                 {"code", "NetStream.Play.Start" },
                 {"description", "Started playing." },
                 {"details", streamName }
-            });
+            }, _streamId, COMMAND_CHANNEL);
             _sessionStorage.ConnectedSession = _publisherSessionService.FindPublisher(streamName);
             if (_sessionStorage.ConnectedSession == null)
             {
@@ -163,31 +168,78 @@ namespace RtmpSharp.Controller
             {
                 throw new InvalidOperationException();
             }
-            _sessionStorage.AudioBuffer = new Queue<AudioData>(publisherSessionStorage.AudioBuffer);
-            _sessionStorage.VideoBuffer = new Queue<VideoData>(publisherSessionStorage.VideoBuffer);
-            publisherSessionStorage.AudioReceived += SendAudioData;
-            publisherSessionStorage.VideoReceived += SendVideoData;
+            _sessionStorage.AudioBuffer = new Queue<AudioData>();
+            _sessionStorage.VideoBuffer = new Queue<VideoData>();
+            publisherSessionStorage.AudioReceived += BufferAudioData;
+            publisherSessionStorage.VideoReceived += BufferVideoData;
+            //ServePlay();
+        }
+        private SemaphoreSlim audioReceived = new SemaphoreSlim(0);
+        private SemaphoreSlim videoReceived = new SemaphoreSlim(0);
+
+        private void ServePlay()
+        {
+            var tsk = SendAvData();
+            tsk.ContinueWith(t =>
+            {
+                if (!_sessionStorage.AudioBuffer.Any() && _sessionStorage.VideoBuffer.Any())
+                {
+                    Session.WriteProtocolControlMessage(new UserControlMessage(UserControlMessageType.StreamDry, new int[] { _streamId }));
+                }
+                ServePlay();
+            }, TaskContinuationOptions.OnlyOnRanToCompletion);
         }
 
-        private void SendAudioData(object sender, AudioEventArgs e)
+        private async Task SendAvData()
         {
+            await audioReceived.WaitAsync();
+            await videoReceived.WaitAsync();
+            Session.SendAmf0Data(_sessionStorage.AudioBuffer.Dequeue(), _streamId, AUDIO_CHANNEL);
+            Session.SendAmf0Data(_sessionStorage.VideoBuffer.Dequeue(), _streamId, VIDEO_CHANNEL);
+        }
+
+        private bool readyForPlay = false;
+        private int videoTimestamp = 0;
+        private int audioTimestamp = 0;
+        public override List<int> CreatedStreams { get; } = new List<int>();
+
+        private void BufferAudioData(object sender, AudioEventArgs e)
+        {
+            if (!readyForPlay)
+            {
+                return;
+            }
+            // audioTimestamp += e.AudioData.Timestamp;
+            // e.AudioData.Timestamp = audioTimestamp;
+            Session.SendAmf0Data(e.AudioData, _streamId, AUDIO_CHANNEL);
+
             // _sessionStorage.AudioBuffer.Enqueue(e.AudioData);
-            // var flvMetadata = (Dictionary<string, object>)_sessionStorage.ConnectedSession.SessionStorage.FlvMetaData.MethodCall.Parameters[0];
-            // var frameCount = Math.Max(1, Session.BufferMilliseconds * ((double)flvMetadata["framerate"] / 1000));
-            // while (_sessionStorage.AudioBuffer.Count >= frameCount)
-            // {
-            Session.SendAmf0Data(e.AudioData);
-            // }
+            // audioReceived.Release();
         }
-
-        private void SendVideoData(object sender, VideoEventArgs e)
+        private void BufferVideoData(object sender, VideoEventArgs e)
         {
+            if (!readyForPlay && e.VideoData.Data[0] >> 4 == 0x01)
+            {
+                Session.WriteProtocolControlMessage(new UserControlMessage(UserControlMessageType.StreamBegin, new int[] { _streamId }));
+                readyForPlay = true;
+            }
+            if (!readyForPlay)
+            {
+                return;
+            }
+
+            // _sessionStorage.VideoBuffer.Enqueue(e.VideoData);
+            // videoReceived.Release();
             // _sessionStorage.VideoBuffer.Enqueue(e.VideoData);
             // var flvMetadata = (Dictionary<string, object>)_sessionStorage.ConnectedSession.SessionStorage.FlvMetaData.MethodCall.Parameters[0];
             // var frameCount = Math.Max(1, Session.BufferMilliseconds * ((double)flvMetadata["framerate"] / 1000));
             // while (_sessionStorage.VideoBuffer.Count >= frameCount)
             // {
-            Session.SendAmf0Data(e.VideoData);
+
+            // videoTimestamp += e.VideoData.Timestamp;
+            // e.VideoData.Timestamp = videoTimestamp;
+
+            Session.SendAmf0Data(e.VideoData, _streamId, VIDEO_CHANNEL);
             // }
         }
 
@@ -216,15 +268,15 @@ namespace RtmpSharp.Controller
                 headerBuffer[8] = dataOffset[0];
                 Session.SendRawData(headerBuffer);
             }
-            Session.SendAmf0Data(_sessionStorage.ConnectedSession.SessionStorage.FlvMetaData);
-            if (hasAudio) Session.SendAmf0Data(_sessionStorage.ConnectedSession.SessionStorage.AACConfigureRecord);
-            if (hasVideo) Session.SendAmf0Data(_sessionStorage.ConnectedSession.SessionStorage.AVCConfigureRecord);
+            Session.SendAmf0Data(_sessionStorage.ConnectedSession.SessionStorage.FlvMetaData, _streamId, COMMAND_CHANNEL);
+            if (hasAudio) Session.SendAmf0Data(_sessionStorage.ConnectedSession.SessionStorage.AACConfigureRecord, _streamId, COMMAND_CHANNEL);
+            if (hasVideo) Session.SendAmf0Data(_sessionStorage.ConnectedSession.SessionStorage.AVCConfigureRecord, _streamId, COMMAND_CHANNEL);
 
         }
 
-        internal override void OnAudio(AudioData audioData)
+        public override void OnAudio(AudioData audioData)
         {
-            if (_sessionStorage.AACConfigureRecord == null && audioData.Data.Length >= 2 && audioData.Data[1] == 0)
+            if (_sessionStorage.AACConfigureRecord == null && audioData.Data.Length >= 2)
             {
                 _sessionStorage.AACConfigureRecord = audioData;
                 return;
@@ -232,10 +284,9 @@ namespace RtmpSharp.Controller
             _sessionStorage.TriggerAudioReceived(this, new AudioEventArgs(audioData));
         }
 
-        internal override void OnVideo(VideoData videoData)
+        public override void OnVideo(VideoData videoData)
         {
-
-            if (_sessionStorage.AVCConfigureRecord == null && videoData.Data.Length >= 2 && videoData.Data[1] == 0)
+            if (_sessionStorage.AVCConfigureRecord == null && videoData.Data.Length >= 2)
             {
                 _sessionStorage.AVCConfigureRecord = videoData;
                 return;
@@ -243,13 +294,31 @@ namespace RtmpSharp.Controller
             _sessionStorage.TriggerVideoReceived(this, new VideoEventArgs(videoData));
         }
 
-        internal override void EnsureSessionStorage()
+        public override void EnsureSessionStorage()
         {
             if (Session.SessionStorage == null)
             {
                 Session.SessionStorage = new SessionStorage();
             }
             _sessionStorage = Session.SessionStorage;
+        }
+
+        [RpcMethod(Name = "createStream", ChannelId = COMMAND_CHANNEL)]
+        public override async Task<int> CreateStream()
+        {
+            if (_streamId != 0)
+            {
+                throw new InvalidOperationException("this controller only supprot one stream");
+            }
+            _streamId = _random.Next(1000, 4000);
+            CreatedStreams.Add(_streamId);
+            return _streamId;
+        }
+
+        public override void DeleteStream()
+        {
+            CreatedStreams.Remove(_streamId);
+            _streamId = 0;
         }
     }
 }
